Fix client output
git-svn-id: file:///.data/svn_convert/svn@141 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/sim/msgedit.cpp
@@ -1580,7 +1580,7 @@ QString MsgEdit::trim(const QString &s)
 
 QString MsgEdit::chunk(const QString &s, int len)
 {
-    if (s.length() < len) return s;
+    if ((int)s.length() < len) return s;
     QString res = s.left(len+1);
     int n = res.length() - 1;
     for (n = res.length() - 1; n >= 0; n--)
